修复大量BUG,优化多处逻辑，添加拖放处理。 修复：分层窗口下无法最大化的问题 优化：优化重绘逻辑，大幅提升了效率（好吧，是因为以前每一次重绘都需要重画一次窗口背景），优化组件自动重绘逻辑,优化组件移动逻辑 新增窗口启动动画，新增图片框，新增用户接入消息管理，新增拖放，并加入获取拖放文件图标的演示。将公用的定义移动至Def.h。新增随鼠标移动属性

+++ DUI_Demo/DUI_Prompt.cpp
@@ -44,10 +44,10 @@ BOOL DUI_Prompt::Create(HWND hParent, LPTSTR Text, RectF* Rect)
 	Info->hinst = GetModuleHandle(nullptr);
 	Info->uId = NULL;
 	Info->lpszText = Text;
-	Info->rect.left = m_Rect->X;
-	Info->rect.top = m_Rect->Y;
-	Info->rect.right = m_Rect->X + m_Rect->Width;
-	Info->rect.bottom = m_Rect->Y + m_Rect->Height;
+	Info->rect.left = (LONG)m_Rect->X;
+	Info->rect.top = (LONG)m_Rect->Y;
+	Info->rect.right = (LONG)m_Rect->X + (LONG)m_Rect->Width;
+	Info->rect.bottom = (LONG)m_Rect->Y + (LONG)m_Rect->Height;
 	SendMessage(m_hWnd, TTM_ADDTOOL, NULL, (LPARAM)Info);
 	SendMessage(m_hWnd, TTM_SETTIPBKCOLOR, RGB(255, 255, 191), (LPARAM)Info);
 	SendMessage(m_hWnd, TTM_SETTIPTEXTCOLOR, 0, (LPARAM)Info);
